Initial skelton for the plugin
I am not clear whether to use a fork, or to use a separate thread to run
the lua script, so kind of stuck at this point. Also using a readpipe is
also pointless at this point as the headers can only be sent only after
the whole script has been executed. Currently thinking that I should
execute the lua script in it's own thread and wake up the socket only
after this is done. This would mean I would have to store the entire
body of the response before writing it.